migtd: public `read_mig_info` and move it to `data` module
